% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastMatrix.R
\name{build.M.c}
\alias{build.M.c}
\title{build.M.c}
\usage{
build.M.c(n)
}
\arguments{
\item{n}{number of nodes for the complete graph}
}
\value{
Returns the contrast matrix of a complete graph with n nodes
}
\description{
Build the contrast matrix based on a complete graph with n nodes
}
\examples{


build.M.c(10)

}
