% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastMatrix.R
\name{build.M.mst}
\alias{build.M.mst}
\title{build.M.mst}
\usage{
build.M.mst(Xj)
}
\arguments{
\item{Xj}{the threshold variable}
}
\value{
Returns the contrast matrix of a MST constructed based on the
threshold variable provided
}
\description{
Build the contrast matrix of an MST constructed based on the threshold
variable provided.
}
\examples{

set.seed(121)
n <- 20
Xj <- runif(n, -1 ,1)

build.M.mst(Xj)

}
