% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastMatrix.R
\name{build.M.nbr}
\alias{build.M.nbr}
\title{build.M.nbr}
\usage{
build.M.nbr(K, nbr.K = NULL, D = NULL)
}
\arguments{
\item{K}{number of neighbors for each node}

\item{nbr.K}{nbr.K[i,] = the node indices ranked by proximity to node i
(from the nearest to the farthest)}

\item{D}{the distance matrix}
}
\value{
Returns the contrast matrix of a KNN graph with n nodes
}
\description{
Build the contrast matrix of an KNN graph with n nodes
}
\examples{


set.seed(121)

n <- 5 # number of nodes
X <- runif(n, -1, 1) # nodes are generated from U(-1, 1)
D <- as.matrix(dist(X)) # distance matrix for the nodes
nbr.K <- t(sortIndex(D)) # ranked indices based on distance
M1 <- build.M.nbr(K = 2, nbr.K = nbr.K[,-1])
M2 <- build.M.nbr(K = 2, D = D)

mean(M1 == M2) == 1 #returns TRUE

}
