% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schlittgen.R
\name{clusterWise}
\alias{clusterWise}
\title{clusterWise}
\usage{
clusterWise(x, y, k, rep = 30)
}
\arguments{
\item{x}{the design matrix (intercept term should be included if any)}

\item{y}{the response vector}

\item{k}{number of clusters}

\item{rep}{number of replications; default  = 30}
}
\value{
Returns the membership labels
}
\description{
build the knn indices based on schlittgens' algorithm
}
