% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.R
\name{getAST}
\alias{getAST}
\title{Recursively constructs Abstract Syntax Tree for a given expression}
\source{
\url{https://stackoverflow.com/a/52080518/6873133}
}
\usage{
getAST(ee)
}
\arguments{
\item{ee}{expression}
}
\value{
list
}
\description{
Recursively constructs Abstract Syntax Tree for a given expression
}
\examples{
getAST(quote(a \%>\% b))
}
