% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.R
\name{h}
\alias{h}
\title{Pipe-able wrapper function around help()}
\source{
\url{https://stackoverflow.com/a/52080518/6873133}
}
\usage{
h(x)
}
\arguments{
\item{x}{expression}
}
\value{
None
}
\description{
This function allows you to pipe forward functions to open its help page.
}
\details{
Using the pipe operator is commonplace and can influence the way you think
about using functions. When looking up help pages, it is typical to add a
question mark \verb{?} to the beginning of a function open its help page (e.g.,
\code{?dim}). This function allows you to write left-to-right and simply pipe the
function of interest into \code{h} to open its help page.
}
\examples{
# Simple single function
eval \%>\% h

# Simple single function with parentheses
dim() \%>\% h

# Function but specifying package
dplyr::across \%>\% h

# Function but specifying package with parentheses
dplyr::summarise() \%>\% h

# Fails
iris \%>\% dplyr::mutate \%>\% h
iris \%>\% dplyr::mutate() \%>\% h
}
