% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{L0}
\alias{L0}
\title{L0
calculate the L0 norm of a vector}
\usage{
L0(v)
}
\arguments{
\item{v}{the vector of which the L0 norm is calculated}
}
\value{
the L0 norm of the vector v
}
\description{
L0
calculate the L0 norm of a vector
}
