% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{L2}
\alias{L2}
\title{L2
calculate the L2 norm of a vector}
\usage{
L2(v)
}
\arguments{
\item{v}{the vector of which the L2 norm is calculated}
}
\value{
the L2 norm of the vector v
}
\description{
L2
calculate the L2 norm of a vector
}
