% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schlittgen.R
\name{sch}
\alias{sch}
\title{sch}
\usage{
sch(X, y, rs, k)
}
\arguments{
\item{X}{the design matrix}

\item{y}{the response vector}

\item{rs}{number of repetitions}

\item{k}{number of clusters}
}
\value{
Returns an n by k matrix; each row a sorted data indices
}
\description{
build the knn indices based on schlittgens' algorithm
}
