% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{softThresh}
\alias{softThresh}
\title{softThresh
returns the soft threshold of a vector}
\usage{
softThresh(x, a)
}
\arguments{
\item{x}{the vector on which the soft threshold operation is performed}

\item{a}{the threshold value}
}
\value{
the soft thresholding of vector x
}
\description{
softThresh
returns the soft threshold of a vector
}
