% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sortIndex}
\alias{sortIndex}
\title{sortIndex}
\usage{
sortIndex(X, decrease = F)
}
\arguments{
\item{X}{a matrix/vector to be sorted}

\item{decrease}{sorted in ascending or descending order, default is FALSE}
}
\value{
the sorted indices for each column vector
}
\description{
sort each column of the given matrix X (default in the ascending order)
}
\examples{


sortIndex(c(2,-1,0), decrease = TRUE)
sortIndex(cbind(c(2,3,1), c(5,10,-1)))


}
