% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.cwr}
\alias{summary.cwr}
\title{summary.cwr}
\usage{
\method{summary}{cwr}(object, ...)
}
\arguments{
\item{object}{a list of lm objects}

\item{...}{further arugments to be passed to}
}
\value{
a list of length K, where K is the number of clusters and each list element is an OLS summary of a subgroup model
}
\description{
OLS summary of the clusterwise regression (cwr) models
}
